fix: 指纹删除多次删除，点击授权框取消，会出现两次授权框
修复指纹删除多次删除，点击授权框取消，会出现两次授权框

Log: 修复了指纹删除多次删除，点击授权框取消，会出现两次授权框
Task: https://pms.uniontech.com/zentao/task-view-23562.html

+++ src/frame/modules/accounts/fingerworker.cpp
@@ -119,10 +119,14 @@ void FingerWorker::stopEnroll(const QString& userName)
     }
 }
 
-void FingerWorker::deleteFingerItem(const QString& userName, const QString& finger)
+void FingerWorker::deleteFingerItem(const QString &userName, const QString &finger)
 {
+    if (m_isDelete)
+        return ;
+    m_isDelete = true;
     auto call = m_fingerPrintInter->DeleteFinger(userName, finger);
     call.waitForFinished();
+    QTimer::singleShot(200, [ = ] () {m_isDelete = false;});
     if (call.isError()) {
         qDebug() << "call DeleteFinger Error : " << call.error();
     }
